[refactor] store Datum.rssi on 8 bits (instead of 32)

+++ src/structs/data.rs
@@ -6,7 +6,7 @@ pub struct Datum {
     name: String,
     sender_mac: String,
     receiver_mac: String,
-    rssi: i32,
+    rssi: i8,
     timestamp: u64
 }
